pyproject.toml
[project]
name = "wealthadvisorai"
version = "0.1.0"
description = "AI-powered multi-agent wealth advisory platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "huggingface-hub>=1.10.1",
    "langchain>=1.2.15",
    "langchain-core>=1.2.28",
    "langchain-huggingface>=1.2.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
